+++ app/controllers/boats_controller.ts
@@ -5,6 +5,7 @@ import AuditLogService from '#services/audit_log_service'
 import BoatListService from '#services/boat_list_service'
 import { toEditForm, toShowProps } from '#transformers/boat_transformer'
 import { toPortFormOptions } from '#transformers/port_transformer'
+import BoatIncidentService from '#services/boat_incident_service'
 import BoatMaintenanceService from '#services/boat_maintenance_service'
 import BoatMaintenanceSheetService from '#services/boat_maintenance_sheet_service'
 import BoatMaintenanceTaskService from '#services/boat_maintenance_task_service'
@@ -27,6 +28,7 @@ export default class BoatsController {
     private maintenanceService: BoatMaintenanceService,
     private taskService: BoatMaintenanceTaskService,
     private sheetService: BoatMaintenanceSheetService,
+    private incidentService: BoatIncidentService,
     private mediaService: MediaService,
     private aiAnalysisService: AiAnalysisService,
     private boatListService: BoatListService,
@@ -117,6 +119,7 @@ export default class BoatsController {
         maintenanceEvents,
         maintenanceTasks,
         maintenanceSheets,
+        incidents,
         boatMedia,
         positionHistory,
         latestSuggestions,
@@ -125,6 +128,7 @@ export default class BoatsController {
         this.maintenanceService.listForBoat(user, boat),
         this.taskService.listForBoat(user, boat),
         this.sheetService.listForBoat(user, boat),
+        this.incidentService.listForBoat(user, boat),
         this.mediaService.listForEntity('boat', boat.id),
         this.boatService.getPositionHistory(boat.id),
         this.aiAnalysisService.getLatestBoatSuggestions(user.id, boat.id),
@@ -144,6 +148,7 @@ export default class BoatsController {
           maintenanceEvents,
           maintenanceTasks,
           maintenanceSheets,
+          incidents,
           aiSuggestions,
           canManageMaintenance,
           canManageEquipment,