+++ app/exceptions/incident_errors.ts
@@ -1,0 +1,13 @@
+export class BoatIncidentNotFoundError extends Error {
+  name = 'BoatIncidentNotFoundError'
+}
+
+export class BoatIncidentValidationError extends Error {
+  name = 'BoatIncidentValidationError'
+  constructor(
+    message: string,
+    readonly errorCode: string
+  ) {
+    super(message)
+  }
+}